Celery setup. PMT #105639
This adds the required libraries for celery and sets up a demo task plus
a view to test it.

The broker settings for staging/prod have been set in
`local_settings.py` and salted out to the servers and the required
rabbitmq vhosts set up.

The only part missing is that there isn't a celery worker running
anywhere yet, but of course, the libraries and such added here need to
be in the code before that can happen.

If this is merged and deployed to staging you ought to be able to test
by:

* log into the staging server and manually start up a worker

    cd /var/www/footprints/footprints/; ./manage.py celeryd \
    --settings=footprints.settings_staging

* then just hit
  https://footprints.staging.ccnmtl.columbia.edu/celerytest/

And you should see it print in the worker console/logs.

Same thing for production. Then I can set up a deployment for a
dedicated worker process and we should be good to go.