main: reduce test setup
Once the postgis code was added, the unit tests began experiencing a segmentation fault.
A complex and heavy usage of the sqlite3 savepoint logic seems to be in play.
Reducing the amount of test setup required alleviates the problem.
Checking this in as a separate PR from the postgis to allow continued experimentation.