Replace 'stringify_role_actors'
This commit replaces 'stringify_role_actors' with
'interpolate_role_actors'. This was done so that the function could
return a list with actors and thier VIAF numbers in alternating fields.
This was an easier method of appending the info to a row as a string,
which was getting munged into a single field.