Remove librabbitmq dependency
Based on a few things I've looked at and local tests, I don't think librabbitmq is a necessary dependency as the amqp library does the same thing. It has stopped compiling on MacOS and the maintainers are encouraging folks to switch. I'm going to pull it, then verify functionality in production.