Switch back to sqlite3
The migration 0019_auto_20151204_0942.py is blowing up in sqlite3/spatialite as
"Renaming table while in a transaction is not supported on SQLite < 3.26 because it
would break referential integrity." This seems to be rooted in the latest version of Django +
sqlite3 + spatialite. A fix theoretically is coming.

(The suggested fix of adding `atomic = False` results in the table *not* being renamed, so things blow up later down the line.)

I'm taking the extreme step here of changing the migration files and removing the rename. While not ideal,
this does resolve the issue and should have no effect in production or our dev environments.