Place: add unique constraint to prevent duplicates
IMPORTANT: Please review & confirm this pull request, but DO NOT MERGE.

This pull request adds a unique constraint on Place alternate_name and canonical_place. It is the last step in a lengthy cleanup process. Before this can be merged, the "consolidate" jobs need to be run in both staging and production to ensure the constraint can be applied.