Improve DigitalObject alt text
This PR adds an alt_text property to Digital Image. Attribute defaults to the "name" field unless "alt_text" is populated. Swapping out less than optimal "image" alt text specification for this value.

This is part 1 of a two part change. Part 2 will add the ability to add custom alt text to the image when image is added or edited.

- Added field to DigitalObject model
- Added format method to retrieve name or alt_text
- Incorporated into serializers and templates to render